+++ packages/fluentui/react-northstar/src/utils/positioner/Popper.tsx
@@ -4,36 +4,26 @@ import * as PopperJs from '@popperjs/core';
 import * as _ from 'lodash';
 import * as React from 'react';
 
+import { getReactFiberFromNode } from '../getReactFiberFromNode';
 import { isBrowser } from '../isBrowser';
 import { getBoundary } from './getBoundary';
 import { getScrollParent } from './getScrollParent';
 import { getPlacement, applyRtlToOffset } from './positioningHelper';
 import { PopperModifiers, PopperProps, PopperPositionFix, PopperJsInstance } from './types';
 
-let reactInstanceKey: string;
-
-const getReactInstanceKey = (elm: Node): string => {
-  if (!reactInstanceKey) {
-    for (const k in elm) {
-      if (k.indexOf('__reactInternalInstance$') === 0) {
-        reactInstanceKey = k;
-        break;
-      }
-    }
-  }
-
-  return reactInstanceKey;
-};
-
 const hasAutofocusProp = (node: Node): boolean | undefined => {
   // https://github.com/facebook/react/blob/848bb2426e44606e0a55dfe44c7b3ece33772485/packages/react-dom/src/client/ReactDOMHostConfig.js#L157-L166
-  return (
-    (node.nodeName === 'BUTTON' ||
-      node.nodeName === 'INPUT' ||
-      node.nodeName === 'SELECT' ||
-      node.nodeName === 'TEXTAREA') &&
-    node[getReactInstanceKey(node)].pendingProps.autoFocus
-  );
+  const isAutoFocusableElement =
+    node.nodeName === 'BUTTON' ||
+    node.nodeName === 'INPUT' ||
+    node.nodeName === 'SELECT' ||
+    node.nodeName === 'TEXTAREA';
+
+  if (isAutoFocusableElement) {
+    return getReactFiberFromNode(node).pendingProps.autoFocus;
+  }
+
+  return false;
 };
 
 function hasAutofocusFilter(node: Node) {